fix ADC2 setting
fix ADC2 setting
change library from Adafruit_ILI9341 to Adafruit_ILI9341_STM

+++ GOscillo/pwm.ino
@@ -127,5 +127,5 @@ void pulse_close(void) {
 }
 
 void setduty(void) {
-  Timer2.setCompare(TIMER_CH4, map(duty, 0, 255, 0, count));
+  Timer2.setCompare(TIMER_CH4, map((unsigned short)duty, 0, 255, 0, count));
 }
